component/observe/screen_understand_component/requirements.txt
vllm
pillow
